requirements.txt
streamlit==1.55.0
python-docx==1.2.0
docx2pdf==0.1.8
Pillow==12.1.1
deep-translator==1.11.4
pypdf==6.9.2
